decomposition-of-a-function-into-modules
